hot fix : route api

+++ Darwin.Back/app/models.py
@@ -29,7 +29,7 @@ class Collection(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     idUser = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     timestamp = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    animal_id = db.Column(db.Integer(5), nullable=False)
+    animal_id = db.Column(db.Integer, nullable=False)
     animal_common_name = db.Column(db.String(120), nullable=False)
     animal_scientific_name = db.Column(db.String(120), nullable=False)
     rarity = db.Column(db.String(120), nullable=False)
